fix: PR feedback 수정 #6 - layer architecture & service level refactoring
