+++ src/main/java/com/project/lala/LalaApplication.java
@@ -2,7 +2,9 @@ package com.project.lala;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableAsync;
 
+@EnableAsync
 @SpringBootApplication
 public class LalaApplication {
