+++ src/main/java/com/project/lala/service/EmailService.java
@@ -1,19 +1,28 @@
 package com.project.lala.service;
 
+import java.time.LocalDateTime;
+
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.scheduling.annotation.Async;
-import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.project.lala.dto.SignUpRequest;
+import com.project.lala.entity.EmailAuth;
+import com.project.lala.repository.EmailAuthRepository;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @Service
-@EnableAsync
 @RequiredArgsConstructor
 public class EmailService {
 
 	private final JavaMailSender javaMailSender;
+
+	private final EmailAuthRepository emailAuthRepository;
 
 	@Async
 	public void sendEmail(String email, String authToken) {
@@ -25,5 +34,12 @@ public class EmailService {
 			lala + "/sign/confirm-email?email=" + email + "&authToken=" + authToken);
 
 		javaMailSender.send(mailMessage);
+	}
+
+	@Transactional
+	public void confirmEmail(SignUpRequest signUpRequest) {
+		EmailAuth emailAuth = emailAuthRepository.findValidAuthByEmail(signUpRequest.getEmail(),
+			signUpRequest.getAuthToken(), LocalDateTime.now()).orElseThrow(IllegalArgumentException::new);
+		emailAuth.useToken();
 	}
 }
